backend/requirements.txt
fastapi
uvicorn
anthropic
openai
Jinja2
dotenv
httpx
pytest
sentence-transformers
qdrant-client
psycopg2-binary
pyyaml
sqlalchemy 
aiosqlite 
flake8
python-multipart
pydantic
pydantic[email]